rename Renderer to Printer
